+++ include/pyawaitable/awaitableobject.h
@@ -4,10 +4,10 @@
 #include <Python.h>
 #include <stdbool.h>
 
+#include <pyawaitable/array.h>
+
 typedef int (*awaitcallback)(PyObject *, PyObject *);
 typedef int (*awaitcallback_err)(PyObject *, PyObject *);
-#define CALLBACK_ARRAY_SIZE 128
-#define VALUE_ARRAY_SIZE 32
 
 typedef struct _pyawaitable_callback
 {
@@ -21,31 +21,23 @@ struct _PyAwaitableObject
 {
     PyObject_HEAD
 
-    // Callbacks
-    pyawaitable_callback aw_callbacks[CALLBACK_ARRAY_SIZE];
-    Py_ssize_t aw_callback_index;
+    pyawaitable_array aw_callbacks;
+    pyawaitable_array aw_object_values;
+    pyawaitable_array aw_arbitrary_values;
+    pyawaitable_array aw_integer_values;
 
-    // Stored Values
-    PyObject *aw_values[VALUE_ARRAY_SIZE];
-    Py_ssize_t aw_values_index;
-
-    // Arbitrary Values
-    void *aw_arb_values[VALUE_ARRAY_SIZE];
-    Py_ssize_t aw_arb_values_index;
-
-    // Integer Values
-    long aw_int_values[VALUE_ARRAY_SIZE];
-    Py_ssize_t aw_int_values_index;
-
-    // Awaitable State
+    /* Index of current callback */
     Py_ssize_t aw_state;
+    /* Is the awaitable done? */
     bool aw_done;
+    /* Was the awaitable awaited? */
     bool aw_awaited;
-    bool aw_used;
-
-    // Misc
+    /* Strong reference to the result of the coroutine. */
     PyObject *aw_result;
+    /* Strong reference to the genwrapper. */
     PyObject *aw_gen;
+    /* Set to 1 if the object was cancelled, for introspection against callbacks */
+    int aw_recently_cancelled;
 };
 
 typedef struct _PyAwaitableObject PyAwaitableObject;
@@ -77,10 +69,5 @@ pyawaitable_await_function_impl(
     awaitcallback_err err,
     ...
 );
-
-int
-alloc_awaitable_pool(void);
-void
-dealloc_awaitable_pool(void);
 
 #endif