+++ include/pyawaitable/genwrapper.h
@@ -18,8 +18,7 @@ genwrapper_next(PyObject *self);
 
 int genwrapper_fire_err_callback(
     PyObject *self,
-    PyObject *await,
-    pyawaitable_callback *cb
+    awaitcallback_err err_callback
 );
 
 PyObject *