+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Significantly reduced awaitable object size by dynamically allocating it.
+- Reduced memory footprint by removing preallocated awaitable objects.
+- Objects returned by a PyAwaitable object's `__await__` are now garbage collected (*i.e.*, they don't leak with rare circular references).
+- Removed limit on number of stored callbacks or values.
+- Switched some user-error messages to `RuntimeError` instead of `SystemError`.
+
 ## [1.3.0] - 2024-10-26
 
 - Added support for `async with` via `pyawaitable_async_with`.